Rename spacetime_ewave modules to wave_engine
Rename several spacetime_ewave* modules to wave_engine* across experiments and update imports accordingly. Affected areas include radial_wave, spring_mass (including Euler/Leapfrog/XPBD variants), and experiments m1–m4; launcher and experiment scripts now import wave_engine modules instead of spacetime_ewave. This is a refactor to standardize engine module names without changing runtime initialization (ti.init remains unchanged).
