+++ openwave/xperiments/_archives/spring_mass/spring_leapfrog.py
@@ -21,7 +21,7 @@ from openwave.common import colormap, constants
 from openwave.i_o import render
 
 import openwave.xperiments._archives.spring_mass.spacetime_medium as medium
-import openwave.xperiments._archives.spring_mass.spacetime_ewave_leapfrog as ewave
+import openwave.xperiments._archives.spring_mass.wave_engine_leapfrog as ewave
 
 # Define the architecture to be used by Taichi (GPU vs CPU)
 ti.init(arch=ti.gpu)  # Use GPU if available, else fallback to CPU
